Personal_Blog/requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
python-dotenv==1.0.1
markdown2==2.4.12 